ffmpeg: add swresample and more decoders/parsers to audio-dec variant
Add the AAC decoder + demuxer and the AAC, AC3, FLAC, and MPEG Audio
parsers, since these seem obviously important for audio decoding.

Also add the swresample library, which is quite small and seems
plausibly useful for audio decoding. For example, shairport-sync
requires it for AirPlay 2 support, so this allows changing
shairport-sync's dependency from libffmpeg-full to libffmpeg-audio-dec,
which makes for a significant reduction in image size.

Signed-off-by: Will Mortensen <willmo@gmail.com>